Trait fitness1::cmn::Delegate

pub trait Delegate {
    fn connection_error(&mut self, HttpError) -> Retry { ... }
}

A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

It contains methods to deal with all common issues, as well with the ones related to uploading media

Provided Methods

fn connection_error(&mut self, HttpError) -> Retry

Called whenever there is an HttpError, usually if there are network problems.

Return retry information.